대댓글 기능 구현 완성, 자신의 댓글만 삭제 가능하도록 구현

+++ src/main/java/com/samplesecurity/service/ReplyService.java
@@ -51,55 +51,38 @@ public class ReplyService {
         Member findMember = findMember(authentication);
         Board board = boardRepository.findById(boardId).get();
         Reply parentReply = replyRepository.findById(parentId).get();
-        parentReply.setAnswerNum(parentReply.getAnswerNum() + 1);
 
         //level , reforder  로직 처리
-        System.out.println("reply.getLevel() = " + reply.getLevel());
-        List<Reply> findReplyList = replyRepository.findAllByLevelAndRefAndIdGreaterThan(reply.getLevel(), reply.getRef(), parentId);
-        for (Reply reply1 : findReplyList) {
-            System.out.println("reply1 = " + reply1);
-        }
+        //자신이 들어갈 곳의 밑바닥 구하기
+        int minRefOfBottom = replyRepository.findBottom(reply);
 
-        // 사이에 들어오는 것을 체크 해줘야함
-        // 1-1, 1-2 사이에 1-1-1 이 중간에 들어올 경우
-        //1-1-1의 reforder를 1-2의 것으로 바꾸고 1-2부터 같은 레벨을 가진 reply의 reforder를 모두 +1 해준다.
-
-        if (findReplyList.size() > 0) {
-            log.info("@@");
-            Reply firstReply = findReplyList.get(0);
-
-            reply.setRefOrder(firstReply.getRefOrder());
-
-            for (Reply reply1 : findReplyList) {
-                reply1.setRefOrder(reply1.getRefOrder() + 1);
+        if (minRefOfBottom != -1) {
+            reply.setRefOrder(minRefOfBottom);
+            List<Reply> allReply = replyRepository.findAllByRefOrderGreaterThanEqualAndRef(minRefOfBottom, reply.getRef());
+            for (Reply findReply : allReply) {
+                findReply.setRefOrder(findReply.getRefOrder() + 1);
             }
-
-        }
-        // 1-2 들어올때를 체크해줘야함
-        //1-1, 1-1-1 다음으로 1-2가 올경우 1-1의 1-1 level 이상의  자식댓글을 가진 reply의 댓글 수+1만큼 1-2 refOrder에 더해준다.
-        else {
-            List<Reply> allByLevel = replyRepository.findAllByLevelGreaterThanEqualAndRef(reply.getLevel(), reply.getRef());
-            int answerCount = 0;
-            if (allByLevel.size() > 0) {
-                for (Reply findReply : allByLevel) {
-                    answerCount += findReply.getAnswerNum();
-                }
-                System.out.println("answerCount = " + answerCount);
-                reply.setRefOrder(reply.getRefOrder() + answerCount);
-            } else {
-                reply.setRefOrder(reply.getRefOrder() + 1);
-            }
+        } else {
+            int maxRefOrder = replyRepository.findMaxRefOrder(reply);
+            reply.setRefOrder(maxRefOrder + 1);
         }
 
         reply.setLevel(reply.getLevel() + 1);
         reply.setBoard(board);
         reply.setMember(findMember);
-        log.info("replyService subReply: " + reply);
+        reply.setToMemberName(parentReply.getMember().getName());
+
         replyRepository.save(reply);
     }
 
-    public void delete(Long replyId) {
-        replyRepository.deleteById(replyId);
+    public void delete(Long replyId, Authentication authentication) {
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        Reply reply = replyRepository.findById(replyId).get();
+        if (userDetails.getUsername().equals(reply.getMember().getEmail())) {
+            replyRepository.deleteById(replyId);
+        } else {
+            throw new RuntimeException(("자신의 댓글만 삭제할 수 있습니다."));
+        }
     }
 
     private Member findMember(Authentication authentication) {
